# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Goyais Contracts
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8787
paths:
...
  /v1/projects:
    get:
      summary: List projects
      parameters:
        - in: query
          name: workspace_id
          schema:
            type: string
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/StandardErrorResponse'
...
  /v1/projects/{project_id}:
    delete:
      summary: Delete project
      parameters:
        - $ref: '#/components/parameters/ProjectIdParam'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/StandardErrorResponse'
...
components:
  parameters:
    CursorParam:
      in: query
      name: cursor
      schema:
        type: string
      description: Opaque cursor; current implementation uses numeric offsets.
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size for cursor-based pagination.
...
    ProjectIdParam:
      in: path
      name: project_id
      required: true
      schema:
        type: string
...
  responses:
    StandardErrorResponse:
      description: Standard API error body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardError'
  schemas:
...
    PermissionMode:
      type: string
      enum: [default, acceptEdits, plan, dontAsk, bypassPermissions]
...
    ListEnvelope:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: {}
        next_cursor:
          type: [string, 'null']
...
    StandardError:
      type: object
      required: [code, message, details, trace_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
    Project:
      type: object
      required: [id, workspace_id, name, repo_path, is_git, current_revision, tokens_in_total, tokens_out_total, tokens_total, created_at, updated_at]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        name:
          type: string
        repo_path:
          type: string
        is_git:
          type: boolean
        default_model_config_id:
          type: string
        token_threshold:
          type: integer
          minimum: 1
        default_mode:
          $ref: '#/components/schemas/PermissionMode'
        current_revision:
          type: integer
          format: int64
        tokens_in_total:
          type: integer
          minimum: 0
        tokens_out_total:
          type: integer
          minimum: 0
        tokens_total:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      required: [workspace_id, name, repo_path]
      properties:
        workspace_id:
          type: string
        name:
          type: string
        repo_path:
          type: string
        is_git:
          type: boolean
...